+++ android/filament-android/src/main/cpp/Camera.cpp
@@ -102,8 +102,17 @@ Java_com_google_android_filament_Camera_nSetModelMatrix(JNIEnv *env, jclass,
         jlong nativeCamera, jfloatArray in_) {
     Camera* camera = (Camera *) nativeCamera;
     jfloat *in = env->GetFloatArrayElements(in_, NULL);
-    camera->setModelMatrix(*reinterpret_cast<const filament::math::mat4f*>(in));
+    camera->setModelMatrix((math::mat4)*reinterpret_cast<const filament::math::mat4f*>(in));
     env->ReleaseFloatArrayElements(in_, in, JNI_ABORT);
+}
+
+extern "C" JNIEXPORT void JNICALL
+Java_com_google_android_filament_Camera_nSetModelMatrixFp64(JNIEnv *env, jclass,
+        jlong nativeCamera, jdoubleArray in_) {
+    Camera* camera = (Camera *) nativeCamera;
+    jdouble *in = env->GetDoubleArrayElements(in_, NULL);
+    camera->setModelMatrix(*reinterpret_cast<const filament::math::mat4*>(in));
+    env->ReleaseDoubleArrayElements(in_, in, JNI_ABORT);
 }
 
 extern "C" JNIEXPORT void JNICALL
@@ -141,9 +150,19 @@ Java_com_google_android_filament_Camera_nGetModelMatrix(JNIEnv *env, jclass,
         jlong nativeCamera, jfloatArray out_) {
     Camera *camera = (Camera *) nativeCamera;
     jfloat *out = env->GetFloatArrayElements(out_, NULL);
-    const filament::math::mat4f& m = camera->getModelMatrix();
+    const filament::math::mat4f& m = (math::mat4f)camera->getModelMatrix();
     std::copy_n(&m[0][0], 16, out);
     env->ReleaseFloatArrayElements(out_, out, 0);
+}
+
+extern "C" JNIEXPORT void JNICALL
+Java_com_google_android_filament_Camera_nGetModelMatrixFp64(JNIEnv *env, jclass,
+        jlong nativeCamera, jdoubleArray out_) {
+    Camera *camera = (Camera *) nativeCamera;
+    jdouble *out = env->GetDoubleArrayElements(out_, NULL);
+    const filament::math::mat4& m = camera->getModelMatrix();
+    std::copy_n(&m[0][0], 16, out);
+    env->ReleaseDoubleArrayElements(out_, out, 0);
 }
 
 extern "C" JNIEXPORT void JNICALL
@@ -151,9 +170,19 @@ Java_com_google_android_filament_Camera_nGetViewMatrix(JNIEnv *env, jclass, jlon
         jfloatArray out_) {
     Camera *camera = (Camera *) nativeCamera;
     jfloat *out = env->GetFloatArrayElements(out_, NULL);
-    const filament::math::mat4f& m = camera->getViewMatrix();
+    const filament::math::mat4f& m = (math::mat4f)camera->getViewMatrix();
     std::copy_n(&m[0][0], 16, out);
     env->ReleaseFloatArrayElements(out_, out, 0);
+}
+
+extern "C" JNIEXPORT void JNICALL
+Java_com_google_android_filament_Camera_nGetViewMatrixFp64(JNIEnv *env, jclass, jlong nativeCamera,
+        jdoubleArray out_) {
+    Camera *camera = (Camera *) nativeCamera;
+    jdouble *out = env->GetDoubleArrayElements(out_, NULL);
+    const filament::math::mat4& m = camera->getViewMatrix();
+    std::copy_n(&m[0][0], 16, out);
+    env->ReleaseDoubleArrayElements(out_, out, 0);
 }
 
 extern "C" JNIEXPORT void JNICALL